Fix C4267 in t5_whisper_medical line 331
Cast inputs[i]->size() / 768 to unsigned int when assigning to
AILIAShape::y to silence C4267.

+++ natural_language_processing/t5_whisper_medical/t5_whisper_medical.cpp
@@ -328,7 +328,7 @@ int forward_decoder(AILIANetwork *ailia, std::vector<float> *inputs[NUM_INPUTS_D
 		if (i == 1){
 			// embeddings
 			sequence_shape.x=768;
-			sequence_shape.y=inputs[i]->size() / 768;
+			sequence_shape.y=(unsigned int)(inputs[i]->size() / 768);
 			sequence_shape.z=batch_size;
 			sequence_shape.w=1;
 			sequence_shape.dim=3;
